+++ inc/frontend/frontend-functions.php
@@ -9,5 +9,3 @@
  */
 
 defined( 'ABSPATH' ) || exit;
-
-