+++ wolf-core.php
@@ -3,7 +3,7 @@
  * Plugin Name: Wolf Core
  * Plugin URI: https://wlfthm.es/wolf-core
  * Description: Core functions for Wolf Themes.
- * Version: 1.8.3
+ * Version: 1.8.5
  * Author: WolfThemes
  * Author URI: https://wolfthemes.com
  * Requires at least: 5.0
@@ -30,7 +30,7 @@ if ( ! class_exists( 'Wolf_Core' ) ) {
 	 * Contains the main functions for Wolf_Core
 	 *
 	 * @class Wolf_Core
-	 * @version 1.8.3
+	 * @version 1.8.5
 	 * @since 1.0.0
 	 */
 	final class Wolf_Core {
@@ -42,7 +42,7 @@ if ( ! class_exists( 'Wolf_Core' ) ) {
 		 *
 		 * @var string The plugin version.
 		 */
-		const VERSION = '1.0.0';
+		const VERSION = '1.8.5';
 
 		/**
 		 * Minimum Elementor Version
@@ -74,7 +74,7 @@ if ( ! class_exists( 'Wolf_Core' ) ) {
 		/**
 		 * @var string
 		 */
-		public $version = '1.8.3';
+		public $version = '1.8.5';
 
 
 
@@ -613,7 +613,7 @@ if ( ! class_exists( 'Wolf_Core' ) ) {
 				// 'template-functions',
 				// 'template-hooks',
 				'body-classes',
-				// 'frontend-functions',
+				'frontend-functions',
 			);
 
 			foreach ( $frontend_files as $file ) {